refactor: receive a target ref rather than provide a ref
In the previous version, `useConditionalFocus()` provided the user a `React.useRef()` object that
they could then provide as a `ref` prop to a component. Now, `useConditionalFocus()` receives that
ref rather than providing it. This allows the hook to be used on elements that aren't strictly in
the React tree or that aren't within scope.

BREAKING CHANGE: `useConditionalFocus()` returns `undefined` now instead of a providing a
`React.useRef()` object